# lookupd: enhance performance for nodes api