@ploxiln 2018-11-08
I did notice the first version, it was pretty simple ... the latest version fails tests with:

```
panic: assignment to entry in nil map

github.com/nsqio/nsq/nsqlookupd.(*RegistrationDB).AddProducer(0xc420122420, 0x8d7cae, 0x6, 0x8d5d70, 0x0, 0x8d5d70, 0x0, 0xc420123440, 0xc420134900)
	/home/travis/gopath/src/github.com/nsqio/nsq/nsqlookupd/registration_db.go:83 +0x543
github.com/nsqio/nsq/nsqlookupd.(*LookupProtocolV1).IDENTIFY(0xc42000e308, 0xc42012aea0, 0xc420021500, 0xc420121cd0, 0x0, 0x0, 0xc420121cd0, 0xc420134c28, 0x4e433b, 0xc420121cc0, ...)
	/home/travis/gopath/src/github.com/nsqio/nsq/nsqlookupd/lookup_protocol_v1.go:228 +0xb7b
```