@andyxning 2018-11-09
@ploxiln After apply the change. One of the clusters in our production reduce `nodes` api response time from 0m0.527s to 0m0.096s.