@ploxiln 2018-11-10
Looks good.

It's funny that `doNodes()` gets a list of all producers, and for each it gets a list of all topics, and then for each topic it gets all producers, just to get the tombstoned state for the producer+topic it is working on ... maybe this could be refactored more, but I'm admittedly too rusty with this codebase to really analyze this at the moment.

Anyway, not invasive, looks like a good little optimization.